added Weather Repository
